osx: twins return long where the x86 returns rax -- chainwork_cmp said lighter chains were heavier
chainwork_cmp was `int` in the twin while reorg.c (and every other
caller) declares `extern long chainwork_cmp`, as the x86 returns it in
rax. On arm64 an int -1 leaves w0 = 0xffffffff with x0's upper half zero,
so the caller's long read 4294967295: reorg_analyze saw a LIGHTER
competing chain as heavier and would reorg onto it. test_reorg on arm64
failed 7 checks on this alone ("cmp b<a on low limb", "analyze reports a
fork that is NOT heavier", "a lighter competing chain: no action", ...).

The same int-vs-long mismatch, found by sweeping every non-static twin
that returns 32 bits against the callers' extern declarations:
store_append (-1 on failure read as height 4294967295, never < 0),
store_get_at (-1 missing / -3 pruned), store_get_tip, store_init,
store_reload, store_get_file_fd, store_get_sync, store_layout_monotonic,
store_truncate_index_only, store_chainwork_truncate; num3072_is_overflow
and utxo_script_unspendable (0/1 today, aligned anyway). All now return
long, like the x86; callers that declare int still read the low half.

Darwin: the daemon links; test_reorg (arm64 copy, the x86-register
node_sync_multi case compiled out) ALL PASS for the first time (19
failures on this branch earlier today, 0 now), plus test_store,
test_archive_check, test_archive_repair, test_archive_truncate_nonmonotonic,
test_append_unshared_frontier, test_frontier_append, test_utxo_unspendable,
test_muhash, test_lsm_mm_collide.

+++ port/osx/chainwork_twin.c
@@ -6,7 +6,7 @@
  *   void u256_div(u8 q_le[32], const u8 a_le[32], const u8 b_le[32]);
  *   void block_work(u64 work[2], u32 bits);
  *   void chainwork_add(u64 out[2], const u64 a[2], const u64 b[2]);
- *   int  chainwork_cmp(const u64 a[2], const u64 b[2]);
+ *   long chainwork_cmp(const u64 a[2], const u64 b[2]);
  *   int  store_chainwork_init(void *st);
  *   int  store_chainwork_append(void *st, long height, const u64 work[2]);
  *   int  store_chainwork_get_at(void *st, long height, u64 out[2]);
@@ -150,7 +150,13 @@ void chainwork_add(u64 out[2], const u64 a[2], const u64 b[2])
     out[0] = s0; out[1] = s1;
 }
 
-int chainwork_cmp(const u64 a[2], const u64 b[2])
+/* 2026-09-24: LONG, as the x86 returns it in rax and every caller declares
+ * it (reorg.c, tests: `extern long chainwork_cmp`). As `int`, -1 left w0 =
+ * 0xffffffff with x0's upper half zero, so a caller reading the long saw
+ * 4294967295: a LIGHTER candidate compared as heavier and reorg_analyze
+ * would reorg onto it (test_reorg: "cmp b<a" / "a lighter competing chain:
+ * no action" failed on arm64 before this). */
+long chainwork_cmp(const u64 a[2], const u64 b[2])
 {
     if (a[1] != b[1]) return a[1] > b[1] ? 1 : -1;
     if (a[0] != b[0]) return a[0] > b[0] ? 1 : -1;
@@ -217,7 +223,7 @@ long store_chainwork_reload(void *st)
     return n;
 }
 
-int store_chainwork_truncate(void *st, long target_height)
+long store_chainwork_truncate(void *st, long target_height)
 {
     (void)st;
     if (target_height < -1) return -1;
